Added timestamps to be able to filter and merge the nodes based on recency
